refactor: remove test from SolAliasResolveTest
